Refactor: Move S3 download tests to S3Supabase describe block
The S3 download tests were previously defined within the `RSpec.shared_context 'when S3Supabase is set up'` block. This was not the ideal location for these tests, as they are specifically related to the `#download_file` method of the `Timet::S3Supabase` class.

Moving these tests to the correct context improves:
  - Readability: The test organization is now more logical and easier to navigate.
  - Maintainability: Tests are now grouped with the code they're testing, making updates and debugging easier.
  - Correctness: Tests are now run in the correct context, avoiding any unexpected interactions with other tests.

In addition, this commit adds:
  - the `s3_supabase` and `bucket_name` let definitions.
  - move the definition of the `s3_config` to the main describe block.
  - Remove duplicated code.